Windows gcc warnings
```
../../tests/session_fixture.c:266:49: warning: conversion to 'size_t' {aka 'long long unsigned int'} from 'int' may change the sign of the result [-Wsign-conversion]
  266 |                 snprintf(filepath[curpath], len + 1, "%s/%s", p, file);
      |                                             ~~~~^~~
../../tests/session_fixture.c:276:49: warning: conversion to 'size_t' {aka 'long long unsigned int'} from 'int' may change the sign of the result [-Wsign-conversion]
  276 |                 snprintf(filepath[curpath], len + 1, "%s", file);
      |                                             ~~~~^~~
../../tests/session_fixture.c:276:57: error: 'snprintf' output may be truncated before the last format character [-Werror=format-truncation=]
  276 |                 snprintf(filepath[curpath], len + 1, "%s", file);
      |                                                         ^
../../tests/session_fixture.c:276:17: note: 'snprintf' output 1 or more bytes (assuming 2) into a destination of size 1
  276 |                 snprintf(filepath[curpath], len + 1, "%s", file);
      |                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../../tests/session_fixture.c:266:57: error: '/' directive output may be truncated writing 1 byte into a region of size between 0 and 1 [-Werror=format-truncation=]
  266 |                 snprintf(filepath[curpath], len + 1, "%s/%s", p, file);
      |                                                         ^
../../tests/session_fixture.c:266:54: note: assuming directive output of 1 byte
  266 |                 snprintf(filepath[curpath], len + 1, "%s/%s", p, file);
      |                                                      ^~~~~~~
../../tests/session_fixture.c:266:17: note: 'snprintf' output 2 or more bytes (assuming 4) into a destination of size 1
  266 |                 snprintf(filepath[curpath], len + 1, "%s/%s", p, file);
      |                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
```

+++ tests/session_fixture.c
@@ -263,7 +263,7 @@ char *srcdir_path(const char *file)
             len = snprintf(NULL, 0, "%s/%s", p, file);
             if(len >= 0) {
                 filepath[curpath] = calloc(1, (size_t)len + 1);
-                snprintf(filepath[curpath], len + 1, "%s/%s", p, file);
+                snprintf(filepath[curpath], (size_t)len + 1, "%s/%s", p, file);
             }
             else {
                return NULL;
@@ -273,7 +273,7 @@ char *srcdir_path(const char *file)
             len = snprintf(NULL, 0, "%s", file);
             if(len >= 0) {
                 filepath[curpath] = calloc(1, (size_t)len + 1);
-                snprintf(filepath[curpath], len + 1, "%s", file);
+                snprintf(filepath[curpath], (size_t)len + 1, "%s", file);
             }
             else {
                return NULL;
